feat: reorganizar importaciones y actualizar tipos de campaña en varios archivos
