+++ src/views/campaigns/api/campaign-hooks.ts
@@ -1,7 +1,8 @@
 'use client'
 
 import { useQuery } from '@tanstack/react-query'
-import { CampaignFilter, getCampaignById, getCampaigns } from './campaign-queries'
+import { CampaignFilter } from '../model/campaign-types'
+import { getCampaignById, getCampaigns } from './campaign-queries'
 import { CAMPAIGN_KEYS } from './query-keys'
 
 import { useAuth } from '@clerk/nextjs'
@@ -11,7 +12,6 @@ export function useCampaignList(filter: CampaignFilter = 'all') {
   return useQuery({
     queryKey: [...CAMPAIGN_KEYS.list(), filter, userId],
     queryFn: () => (userId ? getCampaigns(filter, userId) : []),
-    enabled: !!userId,
     staleTime: 1000 * 60 * 30,
   })
 }
@@ -20,7 +20,6 @@ export function useCampaign(id: string) {
   return useQuery({
     queryKey: CAMPAIGN_KEYS.detail(id),
     queryFn: () => getCampaignById(id),
-    enabled: !!id,
     staleTime: 1000 * 60 * 30,
   })
 }
