+++ src/views/campaigns/model/campaign-types.ts
@@ -2,3 +2,4 @@ export type CampaignFormValues = {
   name: string
   location?: string
 }
+export type CampaignFilter = 'all' | 'public' | 'owned'
