+++ src/views/campaigns/ui/creation/index.ts
@@ -1,5 +1,5 @@
+export * from '../../model/campaign-types'
+export * from '../../model/campaignStore'
 export * from './CampaignCreationForm'
 export * from './CampaignCreationProvider'
 export * from './StepControls'
-export * from '../../model/campaign'
-export * from '../../model/campaignStore'
